# #589 - More efficient box approximation of specific set types