Closes #589.

```julia
julia> E = rand(Ellipsoid);
julia> B = rand(Ball2);

# this branch
julia> @btime box_approximation($E)
  74.695 ns (1 allocation: 96 bytes)
Hyperrectangle{Float64,Array{Float64,1},Array{Float64,1}}([0.0, 0.0], [1.4142135623730951, 1.4142135623730951])

julia> @btime box_approximation($B)
  52.039 ns (1 allocation: 96 bytes)
Hyperrectangle{Float64,Array{Float64,1},Array{Float64,1}}([-1.3320384136380545, -0.5810181282958253], [0.3527491162460336, 0.3527491162460336])

# master
julia> @btime box_approximation($E)
  126.373 ns (3 allocations: 224 bytes)
Hyperrectangle{Float64,Array{Float64,1},Array{Float64,1}}([0.0, 0.0], [1.4142135623730951, 1.4142135623730951])

julia> @btime box_approximation($B)
  289.367 ns (7 allocations: 608 bytes)
Hyperrectangle{Float64,Array{Float64,1},Array{Float64,1}}([-1.3320384136380545, -0.5810181282958253], [0.35274911624603367, 0.35274911624603367])
```

The second commit is unrelated but also brings a speedup.

```julia
# this branch
julia> @btime volume($B)
  10.424 ns (0 allocations: 0 bytes)
33.510321638291124

# master
julia> @btime volume($B)
  508.309 ns (9 allocations: 928 bytes)
33.510321638291124
```